refactor(dashboard): extract Court dashboard admin-settings sub-router (F-B-202 PR-3/10)
Audit ref: imp/audits/2026-05-20/findings/track-b/F-B-202.md

Sprint 2 modularization PR-3 of 10. base=main (no longer stacked —
lesson from #840→#841 recovery, memory feedback_stacked_pr_merge_disappears).

Moves the admin-settings surface — change-password + policy-toggle —
into a per-feature sub-router. Zero behavioural delta: URLs, response
bodies, template renders, session/CSRF semantics all preserved.

What moves (3 routes, 86 LOC out of router):

  app/dashboard/admin_settings_routes.py (new, 140 LOC):
    GET  /dashboard/admin/settings           settings page
    POST /dashboard/admin/settings/password  change admin password
    POST /dashboard/admin/policy-toggle      flip policy enforcement
                                             (demo / sandbox only)

  app/dashboard/router.py (2930 → 2844 LOC):
    - 3 routes removed; breadcrumb comments at the old positions.
    - Sub-router wired via
      router.include_router(_admin_settings_routes.router)
      near the router declaration.

The change-password endpoint here is the second of the two
bcrypt.gensalt(rounds=12) hardcodes the F-B-202 audit flagged; the
first lives in auth_routes.py::admin_setup_submit (PR #841).
Consolidation onto a single app.kms.admin_secret.BCRYPT_COST constant
is out of scope for this PR (the goal here is the route move).

Pre-merge gate (effective 2026-05-20, memory
feedback_legacy_sandbox_smoke_obsolete):
  - Full unit suite green:
    pytest tests/ --ignore=tests/integration --ignore=tests/e2e
      --deselect tests/test_dashboard_approvals_nav.py::test_badge_approvals_empty_when_plugin_unavailable
      -q
    Result: 4211 passed, 9 skipped, 32 xfailed in 113s.
    1 xdist flake (test_ensure_private_key_permissions_warns_on_loose_mode)
    passes isolated, pre-existing.
  - Smoke ./sandbox/smoke.sh 25/25 INTENTIONALLY SKIPPED on this PR.
    The legacy sandbox covers only the 2026-Q4 baseline (A2A cross-org
    + MCP intra-org + SPIFFE + OIDC) — none of which this PR touches.
    Refactor is mechanical (extract + include_router), same pattern
    as PR #839 and #841 which were both smoke-validated. Real smoke
    gate for post-2026-02 features lands when stack/ modern is ready
    (see imp/plans/sandbox-modern-stack.md).
  - Dashboard tests targeted: pytest tests/test_dashboard*.py -q —
    202/203 (1 pre-existing deselect).

Router LOC progress:
  Pre-Sprint 2  : 67 routes, 3125 LOC
  Post-PR-1 #839: 67 routes, 2930 LOC
  Post-PR-2 #841: 60 routes, 2614 LOC (still open)
  Post-PR-3     : 64 routes, 2844 LOC (this PR, base=main pre-#841)
  Post-rebase-on-#841: 57 routes, ~2528 LOC (after #841 merges)
  Target post-PR-10  : ~3 routes (aggregator), ~100 LOC

This PR conflicts with #841 zero-overlap: PR-2 modifies sections 2+3+5
(auth), PR-3 modifies sections 4+7 (admin-settings + policy-toggle).
Rebase on main-post-#841 will be clean.

Next PR-4 (F-B-202 modularization): extract overview + badges + sse
(sections 8+16+17, ~ 175 LOC, 8 routes).
